LayoutTests/platform/mac-bigsur-wk2/http/tests/websocket/tests/hybi/send-arraybuffer-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://127.0.0.1:8880/websocket/tests/hybi/check-binary-messages' failed: The operation couldn’t be completed. (kNWErrorDomainPOSIX error 57 - Socket is not connected)
WebSocket: Send ArrayBuffers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS PASS: Message #0.
PASS PASS: Message #1.
PASS PASS: Message #2.
FAIL closeEvent.wasClean should be true. Was false.
PASS successfullyParsed is true

TEST COMPLETE
